Add functionality to add a note to Hive database
This commit adds the ability to add a note to the Hive database using the `addNote` method in the `AddNoteCubit`. The method takes a `NoteModel` object and adds it to the `noteBox` in Hive. If successful, the state is updated to `AddNoteSuccess`, otherwise `AddNoteFailure` is emitted with the error message.